repository cleaning (bis)
